requirements.txt
discord.py==2.3.2
openai==1.12.0
python-dotenv==1.0.1
aiosqlite==0.19.0
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
requests==2.31.0
